// frontend/src/module/prod-and-svc/cloud-host/model/service.ts
export class Services {
    
    region_org_data: {} = {};
    region_options: any [] = new Array<any>();
    region_value: any = '';
    region_mappings: {} = {};

    zone_org_data: {} = {};
    zone_options: any [] = new Array<any>();
    zone_value: any = '';
    zone_mappings: {} = {};

    cpu_org_data: {} = {};
    cpu_options: any [] = new Array<any>();
    cpu_value: any = '';
    cpu_mappings: {} = {};

    mem_org_data: {} = {};
    mem_map_value_list: any [] = new Array<any>();
    mem_options: any [] = new Array<any>();
    mem_options_disp: any [] = new Array<any>();
    mem_value: any = '';
    mem_mappings: {} = {};

    imagetype_org_data: {} = {};
    imagetype_options: any [] = new Array<any>();
    imagetype_value: any = '';
    imagetype_mappings: {} = {};

    ostype_org_data: {} = {};
    ostype_options: any [] = new Array<any>();
    ostype_value: any = '';
    ostype_mappings: {} = {};

    osversion_org_data: {} = {};
    osversion_map_value_list: any [] = new Array<any>();
    osversion_options: any [] = new Array<any>();
    osversion_options_disp: any [] = new Array<any>();
    osversion_value: any = '';
    osversion_mappings: {} = {};

    storage_org_data: {} = {};
    storage_options: any [] = new Array<any>();
    storage_value: any = '';
    storage_mappings: {} = {};

    timeline_org_data: {} = {};
    timeline_options: any [] = new Array<any>();
    timeline_value: any = '';
    timeline_mappings: {} = {};
}