// frontend/src/module/prod-and-svc/cloud-host/model/order.ts
export class Order {
    
    region: {} = {};
    zone: {} = {};
    cpu: {'id':''} = {'id':''};
    mem: {} = {};
    imagetype: {} = {};
    ostype: {'id':''} = {'id':''};
    osversion: {} = {};
    storage: {'id':''} = {'id':''};
    timeline: {'id':''} = {'id':''};
    count: number = 1;
    totalPrice: number = 0;
}